Level,DC,MC,SC,AC,MAC,HANDWEIGHT,ITEMWEIGHT
Updated bytes to UShorts, also fixed an AttackSpeed hover item buh on
items when attack speed was a negative value.

Player > Level,DC,MC,SC,AC,MAC,WearWeight,HandWeight

Mob > Level, DC,MC,SC,AC,MAC

+++ Server/MirObjects/Monsters/HumanAssassin.cs
@@ -48,12 +48,12 @@ namespace Server.MirObjects.Monsters
             RefreshBase();
 
             MaxHP = (ushort)Math.Min(ushort.MaxValue, MaxHP + PetLevel * 20);
-            MinAC = (byte)Math.Min(byte.MaxValue, MinAC + PetLevel * 2);
-            MaxAC = (byte)Math.Min(byte.MaxValue, MaxAC + PetLevel * 2);
-            MinMAC = (byte)Math.Min(byte.MaxValue, MinMAC + PetLevel * 2);
-            MaxMAC = (byte)Math.Min(byte.MaxValue, MaxMAC + PetLevel * 2);
-            MinDC = (byte)Math.Min(byte.MaxValue, MinDC + PetLevel);
-            MaxDC = (byte)Math.Min(byte.MaxValue, MaxDC + PetLevel);
+            MinAC = (ushort)Math.Min(ushort.MaxValue, MinAC + PetLevel * 2);
+            MaxAC = (ushort)Math.Min(ushort.MaxValue, MaxAC + PetLevel * 2);
+            MinMAC = (ushort)Math.Min(ushort.MaxValue, MinMAC + PetLevel * 2);
+            MaxMAC = (ushort)Math.Min(ushort.MaxValue, MaxMAC + PetLevel * 2);
+            MinDC = (ushort)Math.Min(ushort.MaxValue, MinDC + PetLevel);
+            MaxDC = (ushort)Math.Min(ushort.MaxValue, MaxDC + PetLevel);
 
             if (MoveSpeed < 100) MoveSpeed = 100;
             if (AttackSpeed < 100) AttackSpeed = 100;
